requirements.txt
pillow==10.2.0
pandas==2.2.1
streamlit==1.32.0